Replace EV curve tests
